+++ carp_communication_package/lib/communication_probes.dart
@@ -10,8 +10,6 @@ part of communication;
 /// A probe that collects the phone log from this device.
 /// Only collects this information once when the [getDatum] method is called.
 class PhoneLogProbe extends DatumProbe {
-  PhoneLogProbe() : super();
-
   Future<Datum> getDatum() async {
     int days = (measure as PhoneLogMeasure).days;
     Iterable<CallLogEntry> entries;
@@ -39,27 +37,18 @@ class PhoneLogProbe extends DatumProbe {
 /// If you want to listen to text messages being received,
 /// use a [TextMessageProbe] instead.
 class TextMessageLogProbe extends DatumProbe {
-  TextMessageLogProbe() : super();
-
   Future<Datum> getDatum() async {
     SmsQuery query = new SmsQuery();
     List<SmsMessage> _messages = await query.getAllSms;
-    return TextMessageLogDatum()
-      ..textMessageLog = _messages.map(_smsToTextMessage).toList();
+    return TextMessageLogDatum()..textMessageLog = _messages.map(_smsToTextMessage).toList();
   }
 
-  TextMessage _smsToTextMessage(SmsMessage sms) =>
-      TextMessage.fromSmsMessage(sms);
+  TextMessage _smsToTextMessage(SmsMessage sms) => TextMessage.fromSmsMessage(sms);
 }
 
 /// The [TextMessageProbe] listens to SMS messages and collects a
 /// [TextMessageDatum] every time a new SMS message is received.
 class TextMessageProbe extends StreamProbe {
-  SmsReceiver receiver = SmsReceiver();
-
-  TextMessageProbe() : super(textMessageStream);
+  Stream<Datum> get stream =>
+      SmsReceiver().onSmsReceived.map((event) => TextMessageDatum.fromTextMessage(TextMessage.fromSmsMessage(event)));
 }
-
-Stream<Datum> get textMessageStream =>
-    SmsReceiver().onSmsReceived.map((event) =>
-        TextMessageDatum.fromTextMessage(TextMessage.fromSmsMessage(event)));