+++ carp_context_package/lib/activity_probe.dart
@@ -10,9 +10,6 @@ part of context;
 /// Collects activity information from the underlying OS's activity recognition API.
 /// It generates an [ActivityDatum] every time an activity is detected.
 class ActivityProbe extends StreamProbe {
-  //ActivityProbe(Measure measure) : super(measure, activityStream);
-  ActivityProbe() : super(_activityStream);
+  Stream<Datum> get stream =>
+      ActivityRecognition.activityUpdates().map((activity) => ActivityDatum.fromActivity(activity));
 }
-
-Stream<Datum> get _activityStream => ActivityRecognition.activityUpdates()
-    .map((activity) => ActivityDatum.fromActivity(activity));