+++ carp_mobile_sensing/lib/sampling_packages/connectivity/connectivity_probes.dart
@@ -10,13 +10,10 @@ part of connectivity;
 /// The [ConnectivityProbe] listens to the connectivity status of the phone and
 /// collect a [ConnectivityDatum] every time the connectivity state changes.
 class ConnectivityProbe extends StreamProbe {
-  //ConnectivityProbe(Measure measure) : super(measure, connectivityStream);
-  ConnectivityProbe() : super(connectivityStream);
+  Stream<Datum> get stream => Connectivity()
+      .onConnectivityChanged
+      .map((ConnectivityResult event) => ConnectivityDatum.fromConnectivityResult(event));
 }
-
-Stream<Datum> get connectivityStream => Connectivity()
-    .onConnectivityChanged
-    .map((ConnectivityResult event) => ConnectivityDatum.fromConnectivityResult(event));
 
 // This probe requests access to location PERMISSIONS (on Android). Don't ask why.....
 // TODO - implement request for getting permission.
@@ -30,20 +27,11 @@ class BluetoothProbe extends PeriodicStreamProbe {
   /// Default timeout for bluetooth scan - 2 secs
   static const DEFAULT_TIMEOUT = 2 * 1000;
 
-//  BluetoothProbe(PeriodicMeasure measure)
-//      : super(
-//            measure,
-//            FlutterBlue.instance
-//                .scan(
-//                    scanMode: ScanMode.lowLatency, timeout: Duration(milliseconds: measure.duration ?? DEFAULT_TIMEOUT))
-//                .map((result) => BluetoothDatum.fromScanResult(result)));
+  Stream<Datum> get stream => FlutterBlue.instance
+      .scan(scanMode: ScanMode.lowLatency, timeout: duration ?? Duration(milliseconds: DEFAULT_TIMEOUT))
+      .map((result) => BluetoothDatum.fromScanResult(result));
 
-  BluetoothProbe()
-      : super(FlutterBlue.instance
-            .scan(scanMode: ScanMode.lowLatency, timeout: Duration(milliseconds: DEFAULT_TIMEOUT))
-            .map((result) => BluetoothDatum.fromScanResult(result)));
-
-  // important that we don't propagate this up the chain, since this would close the overall stream
+  // important to overwrite onDone() as a no-op, so that the super class (StreamProbe) don't close the overall stream.
   @override
   void onDone() {}
 }